Rename FireStoreFileDao to DirectoryDao; make new FileDao for the new file object
